Port src/decomposition. See TODOs
